fix deleteUnverifiedAccount()

+++ mangajutsu-api/src/main/java/com/mangajutsu/api/controllers/UserController.java
@@ -1,5 +1,7 @@
 package com.mangajutsu.api.controllers;
 
+import java.util.List;
+
 import com.mangajutsu.api.dao.entities.UserEntity;
 import com.mangajutsu.api.dao.entities.VerifTokenEntity;
 import com.mangajutsu.api.dao.repositories.UserRepository;
@@ -8,7 +10,6 @@ import com.mangajutsu.api.services.UserService;
 import com.mangajutsu.api.services.VerifTokenService;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Sort;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -28,13 +29,16 @@ public class UserController {
     }
 
     @DeleteMapping("/deleteUnverifiedAccount")
-    public void deleteUnverifiedAccount() {
-        VerifTokenEntity verifToken = verifTokenRepository.findAll(Sort.by(Sort.Direction.DESC,
-                "createdAt")).get(0);
-        if (verifToken.isExpired()) {
-            UserEntity unverifiedAccount = verifToken.getUser();
-            verifTokenService.removeVerifToken(verifToken);
-            userRepository.delete(unverifiedAccount);
+    public void deleteUnverifiedAccount(VerifTokenEntity verifTokenEntity) {
+        if (verifTokenEntity != null) {
+            List<VerifTokenEntity> verifTokens = verifTokenRepository.findAll();
+            for (VerifTokenEntity verifToken : verifTokens) {
+                if (verifToken.isExpired()) {
+                    UserEntity unverifiedAccount = verifToken.getUser();
+                    verifTokenService.removeVerifToken(verifToken);
+                    userRepository.delete(unverifiedAccount);
+                }
+            }
         }
     }
 }
